Sidebar navigation style modifications
- Made links actually a width -.-
- Changed About => Browse
- Update hover, focus, & active styles
- Created 3D effect on nav bar

Up for suggestions on this, I just disliked it before and wanted to get away from programming for a little.

Decided to not have a selected state. Is it necessary? If we have to show what page you are on, are we effectively communicating the content of the page? Look at Amazon, I don't see a selected state. Look at Google, Facebook. You just KNOW where you are. Ideas welcome.